#10949 was squash-merged as 954514c9 and its branch deleted, so GitHub
retargeted this PR onto master. That produced add/add/content conflicts across
the four auto_backport files purely as a history artifact: this branch carries
#10949's seven pre-squash commits, which aren't ancestors of master, so git saw
both sides introducing the same content via unrelated history.

master's copy of all four files is byte-identical to #10949's tip (2df47799),
verified per-file, so every conflict resolves to this branch's version — the
PrivilegedGit consolidation on top of the same base content. Nothing from master
is discarded.

Resulting tree is identical to rebasing this branch's single commit onto master.
73 tests pass.

+++ src/redisearch_rs/query_eval/tests/integration/fuzzy.rs
@@ -32,11 +32,11 @@ use crate::util::{ALL_INDEXED_FIELDS, term_records};
 
 /// The `apiVersion` from which a fuzzy expansion also covers the empty term.
 /// Below it the empty term is never expanded to, whatever the distance.
-const API_VERSION_EMPTY_TERM: u32 = 2;
+const API_VERSION_EMPTY_TERM: u8 = 2;
 
 /// The highest `apiVersion` still below [`API_VERSION_EMPTY_TERM`], so the
 /// negative controls sit right against the gate rather than far below it.
-const API_VERSION_NO_EMPTY_TERM: u32 = API_VERSION_EMPTY_TERM - 1;
+const API_VERSION_NO_EMPTY_TERM: u8 = API_VERSION_EMPTY_TERM - 1;
 
 /// The longest pattern, in runes, the trie's fuzzy walk accepts. A longer one
 /// builds no iterator at all. Read from the C constant rather than restated, so
@@ -112,7 +112,7 @@ struct FuzzyOptions {
     max_expansions: Option<u32>,
     /// The search context's `apiVersion`, which gates whether the expansion also
     /// covers the empty term.
-    api_version: u32,
+    api_version: u8,
 }
 
 impl Default for FuzzyOptions {
